refact: InputVie, ResultView 분리 대응 테스트 구현
https://github.com/next-step/kotlin-racingcar/pull/568#discussion_r873576258
